Rename mavenExplorerProvider.
